Email Verification
Created a system to allow users to verify their
email addresses.  On new-user-create, a confirmation
token will be created and emailed to the user as a link.
Upon visiting that link, their email will be confirmed and
they will be able to login.  Reconfirmation is not required
on email change.